+++ common/token.h
@@ -10,7 +10,7 @@
 #define TOKEN_UNKNOWN 0xFFFF
 
 char token_control[9][MAX_OPCODE_STR_LENGTH];
-char token_memory[5][MAX_OPCODE_STR_LENGTH];
+char token_memory[7][MAX_OPCODE_STR_LENGTH];
 char token_push[2][MAX_OPCODE_STR_LENGTH];