+++ compiler/src/cli.h
@@ -14,10 +14,10 @@
 #define CLI_LABELS  "labels"
 #define CLI_READ "-r"
 
-#define CLI_FLAG_COMPULE "-c"
+#define CLI_FLAG_COMPILE "-c"
 #define CLI_FLAG_HELP "-h"
 #define CLI_FLAG_INTERPRETER "-i"
-#define CLI_FLAG_COMPULE "-c"
+#define CLI_FLAG_COMPILE "-c"